@@ -18,7 +18,7 @@ namespace Raptor.OpenGL
     [ExcludeFromCodeCoverage]
     internal class GameWindowFacade : IGameWindowFacade
     {
-        private readonly object objectLock = new object();
+        private readonly object objectLock = new ();
         private GameWindow? gameWindow;
         private bool isDisposed;
 
@@ -582,7 +582,7 @@ namespace Raptor.OpenGL
             {
                 if (disposing)
                 {
-                    if (!(this.gameWindow is null))
+                    if (this.gameWindow is not null)
                     {
                         this.gameWindow.Dispose();
                     }
